Fix: don't save object status document to galley settings
This commit mostly reverts a previous commit that checked file
deposit status and stored the status details in the galley
settings. This was reverted because it was incorrect. The status
information in a SWORDv3 StatusDocument never pertains to the
file, but always the object. A  in the document may include
a status, but there is no way to connect that to a specific galley.